add active class to filter button
